requirements.txt
alabaster==0.7.12
attrs==19.1.0
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
cliff==2.14.1
cmd2==0.9.11
colorama==0.4.1
configobj==5.0.6
Django==2.0
django-cors-headers==2.4.0
django-extra-fields==1.0.0
django-filter==2.1.0
django-paypal==0.2.5
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
filelock==3.0.10
gnupg==2.3.1
humanfriendly==4.18
idna==2.8
imagesize==1.1.0
Jinja2==2.10
lxml==4.3.2
MarkupSafe==1.1.1
numpy==1.16.1
packaging==19.0
pbr==5.1.3
Pillow==5.4.1
pluggy==0.9.0
prettytable==0.7.2
property-manager==2.3.1
psutil==5.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
pycryptodome==3.7.3
pyflakes==2.1.1
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.3.1
pyperclip==1.7.0
python-secrets==19.3.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
South==1.0.2
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
stevedore==1.30.1
toml==0.10.0
tox==3.7.0
urllib3==1.24.1
verboselogs==1.7
virtualenv==16.4.3
wcwidth==0.1.7
xkcdpass==1.17.2
yamlreader==3.0.4
